+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Data asset schema definition for representing various types of information assets
 - Simple web interface with React and TypeScript
 - Basic visualization of threat model components
 - Navigation between different views